lib: move curl_path.[ch] into vssh/
As this contains code only used by SSH backends.
