+++ lib/vssh/curl_path.c
