+++ lib/vssh/curl_path.h
