+++ tests/unit/unit2604.c
@@ -22,7 +22,7 @@
  *
  ***************************************************************************/
 #include "curlcheck.h"
-#include "curl_path.h"
+#include "vssh/curl_path.h"
 #include "memdebug.h"
 
 static CURLcode unit_setup(void)
